Finished SAMv2 implementation

+++ src/main/kotlin/org/taktik/icure/services/external/rest/v1/dto/be/samv2/embed/QuantityDto.kt
@@ -1,3 +1,5 @@
 package org.taktik.icure.services.external.rest.v1.dto.be.samv2.embed
 
-class QuantityDto(var value: Int? = null, var unit: String? = null)
+import java.math.BigDecimal
+
+class QuantityDto(var value: BigDecimal? = null, var unit: String? = null)
